chore(php) Move to PHP 7.
